Remove asdf plugin from .zshrc because its bad
I swear the asdf plugin for zsh was written by monkeys. They load
the bash script which has unsupported commands. The `complete`
command is not natively in zsh. Why not load in the script specifically
made for zsh??? who knows. I do know I am not using it because it is
stupid and they are stupid for making it. 1/10