Actually fix centaur tab problem
There was a github issue with the actual full solution and it completely
fixes the problem. The problem was in projectile projects, they wouldnt
be tabbed and it woulndt be loaded. There was a string(I believe
commented out in `config.el`) that would work while loaded but not on
init.