Add: starship symbol to show in container
This is for distrobox. The thinking is if I have starship in a shell. I
put effort and care into it and so id know which ones which. I currently
cannot see a situation where I have multiple containers with starship
actively being used at the same time. Anyway so the symbol is enough of
a message to know but also not enough where its annoying and bugs me